Enhanced rmk layer functionality
Added new features and improved existing ones in the rmk layer. Introduced usage of Future for asynchronous operations. Added 'stat' functions to RmkAccessor, allowing it to retrieve file or directory status information. Enhanced the 'list' function to return a list of entries instead of an empty default list. Also, implemented the 'next' function in RmkLister to iterate over entries. In addition, added a 'get' method in InodeTable for retrieving specific entries by path. Lastly, added an 'is_dir' method in RmkMetadata to check if an entry is a directory based on its type.